+++ test/support/fake_pipedrive_server.ex
@@ -22,6 +22,7 @@ defmodule ExPipedrive.FakePipedriveServer do
   import ExPipedrive.FakePersonV2ApiHandler
   import ExPipedrive.FakePipelineApiHandler
   import ExPipedrive.FakePipelineV2ApiHandler
+  import ExPipedrive.FakeStageV2ApiHandler
 
   plug(:match)
 
@@ -195,7 +196,7 @@ defmodule ExPipedrive.FakePipedriveServer do
     |> handle_list_own_activities()
   end
 
-  # --- API v2 (deals + persons + organizations + activities + pipelines) ---
+  # --- API v2 (deals + persons + organizations + activities + pipelines + stages) ---
 
   get "/api/v2/deals" do
     handle_list_deals_v2(conn, conn.query_params)
@@ -291,6 +292,26 @@ defmodule ExPipedrive.FakePipedriveServer do
 
   get "/api/v2/pipelines/:id" do
     handle_get_pipeline_v2(conn, conn.params)
+  end
+
+  get "/api/v2/stages" do
+    handle_list_stages_v2(conn, conn.query_params)
+  end
+
+  post "/api/v2/stages" do
+    handle_create_stage_v2(conn)
+  end
+
+  patch "/api/v2/stages/:id" do
+    handle_update_stage_v2(conn)
+  end
+
+  delete "/api/v2/stages/:id" do
+    handle_delete_stage_v2(conn, conn.params)
+  end
+
+  get "/api/v2/stages/:id" do
+    handle_get_stage_v2(conn, conn.params)
   end
 
   match _ do